Implemented workaround for tick labels being reformated.
This was a result of matplotlib not making tick labels
available *before* a plot is drawn. The solution is to
set the tick labels to the ticks. The problem with this
approach is that since the ticks are floats, that if
they're whole numbers, they will be displayed with
a trailing zero when adding custom ticks.

Workaround is to handle this parsing ourself, rather
than let matplotlib try to guess what we're doing.